[+] functional directory with server functionality
